backend/requirements.txt
# Django and core dependencies
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2

# Database drivers
psycopg2-binary==2.9.7
pymongo==4.6.0

# Authentication
firebase-admin==6.4.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# API and utilities
requests==2.31.0
python-decouple==3.8
celery==5.3.4
redis==5.0.1

# Development and testing
pytest==7.4.3
pytest-django==4.7.0
black==23.11.0
flake8==6.1.0

# Data processing
pandas==2.1.4
numpy==1.25.2
scikit-learn==1.3.2

# File handling
Pillow==10.1.0
python-magic==0.4.27

# Monitoring and logging
sentry-sdk==1.38.0
